feat: Add output that will provide aws-auth data
One of the requirements of EKS is that you create a config map in the
kube-system namespace with a list role ARN's that are allowed to join
the EKS cluster. In order to do this dynamically, we now export the
contect for the current worker pool as an output that will be consumed
by the EKS cluster module to create the configmap.

The output is not of type string but a list of maps that you can pass to
our EKS module.